Fix permissions check for deleting items

+++ backend/src/resolvers/Mutation.js
@@ -45,7 +45,16 @@ const Mutation = {
   },
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id };
-    await ctx.db.query.item({ where }, '{ id title }');
+    const item = await ctx.db.query.item({ where }, '{ id title user { id }}');
+
+    // check if user owns that item or have the right permissions
+    const ownsItem = item.user.id === ctx.request.userId;
+    const hasPermissions = ctx.request.user.permissions.some(permission =>
+      ['ADMIN', 'ITEMDELETE'].includes(permission)
+    );
+    if (!ownsItem && hasPermissions) {
+      throw new Error('You dont have permission to do that!');
+    }
 
     return ctx.db.mutation.deleteItem({ where }, info);
   },
